90% completed
we now have all the internal pieces we need to aminate them and add the loading scene.
