+++ src/components/sections/skills.tsx
@@ -1,8 +1,7 @@
+import type { DataType } from "@/types"
+
 interface Props {
-  data: {
-    title: string
-    items: string[]
-  }[]
+  data: DataType
 }
 
 export const Skills: React.FC<Props> = ({ data }) => {
@@ -12,7 +11,7 @@ export const Skills: React.FC<Props> = ({ data }) => {
         <h3 className="span-w-1 text-end text-[#AFAFB6]"> My skills</h3>
       </div>
       <div className="flex w-1/2 gap-base-2 flex-col sm:flex-row justify-end">
-        {data.map((skills) => (
+        {data.skills.map((skills) => (
           <div key={skills.title} className="span-w-1 flex flex-col gap-2">
             <h3 className="text-[#AFAFB6] text-end">{skills.title}</h3>
             <ul>
