+++ src/pages/index.tsx
@@ -1,18 +1,18 @@
-import React from "react"
-
 import { data } from "@/constants"
 
-import { Footer } from "@/components/sections"
+import { Footer } from "@/components/sections/footer"
 import { Header } from "@/components/sections/header"
 import { Skills } from "@/components/sections/skills"
 import { Works } from "@/components/sections/works"
+import { About } from "@/components/sections/about"
 
 const Index: React.FC = () => {
   return (
     <>
       <Header data={data} />
-      <Works data={data.works} />
-      <Skills data={data.skills} />
+      <About data={data} />
+      <Works data={data} />
+      <Skills data={data} />
       <Footer data={data} />
     </>
   )
