+++ src/types/index.ts
@@ -1,21 +1,27 @@
+export interface WorkType {
+  index: string
+  title: string
+  year: string
+  preview?: string
+  description: string
+  images: string[]
+}
+
+export interface SkillType {
+  title: string
+  items: string[]
+}
+
+export interface SocialType {
+  title: string
+  link: string
+}
+
 export interface DataType {
   name: string
   bio: string
   hero_picture: string
-  works: {
-    index: string
-    title: string
-    year: string
-    preview?: string
-    description: string
-    images: string[]
-  }[]
-  skills: {
-    title: string
-    items: string[]
-  }[]
-  socials: {
-    title: string
-    link: string
-  }[]
+  works: WorkType[]
+  skills: SkillType[]
+  socials: SocialType[]
 }
